Sprit model between squatrt and sit-up Gesture Recognizer module is done! Congratulation.

+++ TeamHyperFit/TeamHyperFit/ViewControllers/TTSessionSummaryViewController.m
@@ -35,7 +35,7 @@
     self.activityNameLabel.text = self.activityName;
     self.activityImageView.image = [UIImage imageNamed:self.activityImageName];
     self.fitpointsLabel.text = [NSString stringWithFormat:@"%ld", self.numberRaps * 100];
-    self.rapsNumberLabel.text = [NSString stringWithFormat:@"%ld", self.numberRaps];
+    self.rapsNumberLabel.text = [NSString stringWithFormat:@"%ld", (long)self.numberRaps];
     
     #warning Please Remove This Test code.
     //! Only for testing.
